fix(module/t2st4): perbaiki TypeError pada komponen datatable modul (OPS-29) v1.0
