+++ features/manage-module/components/ModuleFilter.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen, fireEvent, act } from '@testing-library/react'
+import { render, screen, fireEvent} from '@testing-library/react'
 import { ModuleFilter } from './ModuleFilter'
 import { ModuleStatus } from '../types'
