+++ features/manage-module/components/ModuleTable/index.ts
@@ -1,0 +1,2 @@
+export { DataTable } from './DataTable'
+export { columns } from './columns'
